// src/types/typePlumbingFixtures.ts
export type TypePlumbingFixtures = {
    id: number
    anti_splash: string
    armrest: string
    article: string
    body_coating: string
    body_material: string
    brand: string
    built_in_bidet: string
    category: string
    collection: string
    created_at: string
    depth: string
    dimensions_height: string
    dimensions_length: string
    dimensions_width: string
    discharge_type: string
    door_closers: string
    drawer_closers: string
    drawers_count: string
    facade_material: string
    flush_mode: string
    flush_type: string
    has_drawers: string
    has_shelf: string
    link: string
    main_material: string
    manufacturer: string
    material: string
    mounting_type: string
    name: string
    photos: string
    image1?: string
    price: any
    product_color: string
    product_type: string
    profile_color: string
    rimless: string
    seat_included: string
    seat_lift: string
    seat_material: string
    sink_type: string
    stock: string
    style: string
    tank_volume: string
    updated_at: string
    volume: string
    warranty: string
    washbasin_type: string
    type: string
}